backend/requirements/base.txt
asgiref==3.5.2
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.5
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
environ==1.0
executing==0.8.3
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.2.2.post1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
